Task 2 in "Sorting" compiled and debugged

+++ algorithmization/Array.java
@@ -285,7 +285,7 @@ public class Array {
 		 int i = 0;
 		 while(i < array.length - 1) {
 		 	if(array[i] > array[i+1]) {
-		 		int tmp = array[i];
+		 		double tmp = array[i];
 		 		array[i] = array[i+1];
 		 		array[i+1] = tmp;
